requirements.txt
python-magic==0.4.25
httplib2==0.20.4
apiclient==1.0.4
oauth2client==4.1.3
logbook==1.5.3
progress-bar==8
boto3==1.21.32
google-api-python-client==2.42.0
google-auth-oauthlib==0.5.1
